Fix -Wuseless-cast G++ warnings

+++ src/cpu/cpu.cpp
@@ -609,7 +609,7 @@ byte* CPU::tick() {
         ostringstream oss;
         return_exception = "InstructionUnchanged";
         oss << "instruction pointer did not change, possibly endless loop\n";
-        oss << "note: instruction index was " << (long)(instruction_pointer-bytecode) << " and the opcode was '" << OP_NAMES.at(OPCODE(*instruction_pointer)) << "'";
+        oss << "note: instruction index was " << (instruction_pointer-bytecode) << " and the opcode was '" << OP_NAMES.at(OPCODE(*instruction_pointer)) << "'";
         if (OPCODE(*instruction_pointer) == CALL) {
             oss << '\n';
             oss << "note: this was caused by 'call' opcode immediately calling itself\n"
